add a plane.erase function (to be used from lua)

+++ src/kernel/save.c
@@ -954,7 +954,7 @@ static region *readregion(struct gamedata *data, int x, int y)
             rsetherbtype(r, NULL);
         }
         READ_INT(data->store, &n);
-        rsetherbs(r, (short)n);
+        rsetherbs(r, n);
         READ_INT(data->store, &n);
         if (n < 0) {
             /* bug 2182 */
